pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "get_model"
version = "0.1.0"
description = "GET: a foundation model of transcription across human cell types"
readme = "README.md"
authors = [
    { name = "Xi Fu", email = "fuxialexander@gmail.com" }
]
license = { text = "CC BY-NC 4.0" }
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "einops",
    "hydra-core",
    "lightning",
    "matplotlib",
    "numpy",
    "omegaconf",
    "pandas",
    "pyranges @ git+https://github.com/pyranges/pyranges@master",
    "minlora @ git+https://github.com/cccntu/minLoRA.git@main",
    "scipy",
    "seaborn",
    "timm",
    "torch",
    "torchmetrics",
    "tqdm",
    "gcell @ git+https://github.com/fuxialexander/gcell.git@main",
    "wandb",
    "zarr",
]

[project.urls]
Documentation = "https://fuxialexander.github.io/get_model/"
Repository = "https://github.com/fuxialexander/get_model"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]  # Start looking for packages inside the root folder
exclude = ["test*"]  # Exclude test packages
include = ["get_model*"]  # Only include packages starting with "get_model"

[tool.setuptools.package-data]
"get_model" = ["config/**/*.yaml", "**/*.yaml", "**/*.json", "config/*.yaml"]

[tool.pytest.ini_options]
pythonpath = "./get_model"
addopts = [
    "--import-mode=importlib",
]
log_cli = 1